complete button working

+++ server/public/scripts/client.js
@@ -2,32 +2,6 @@ $(function(){
   console.log("JS fired up!");
   getList();
 
-  //GET List function
-  function getList(){
-    clearDom();
-    $.ajax({
-      type: "GET",
-      url: "/list",
-      success: function(response){
-        console.log("getList called");
-        console.log(response);
-        var list = response;
-        for (var i = 0; i < response.length; i++) {
-          var $el = $('#listContainer').children().last();
-          console.log(list[i].id);
-          console.log(list[i].description);
-          console.log(list[i].completed);
-          $el.append("<div data-completed='"+ list[i].completed + "' data-id='" + list[i].id + "'><p>" + list[i].description + "</p><button class='deleteButton'>Delete</button><button class='completeButton'>Complete</button></div>"
-            // "<div class='listDiv' id='listDiv" + i +
-            //           "'<button class='deleteButton' id='deleteButton" + i +
-            //           "'>Delete</button><button class='completeButton' id='completeButton" + i +
-            //           "'>Complete</button></div"
-                    );
-         } //end for loop
-      }// end success
-    }); //end ajax
-  } //end getList
-
   //POST function to add new list item
   $('#listForm').on('submit', function(event){
     event.preventDefault();
@@ -47,10 +21,13 @@ $(function(){
   //delete button function
   $("#listContainer").on('click', ".deleteButton", function(){
     console.log("delete button clicked");
-    console.log($(this).parent().data('id'));
-    var listId = $(this).parent().data('id');
-    console.log(listId);
-    $.ajax({
+    var confirmDelete = confirm("Delete Note/nAre You Sure?");
+    if (confirmDelete === true) {
+     console.log("You pressed OK!");
+     console.log($(this).parent().data('id'));
+     var listId = $(this).parent().data('id');
+     console.log(listId);
+     $.ajax({
       type: "DELETE",
       url: "list/delete/" + listId + "/",
       success: function (){
@@ -58,6 +35,9 @@ $(function(){
         getList();
       }
     });//end ajax
+  } else {
+    console.log("You pressed Cancel!");
+  }
   });// end delete function
 
     //on click put function in order to change a note to completed
@@ -67,9 +47,33 @@ $(function(){
       console.log($(this).parent().data('completed'));
     });
 
-
-
 }); // end Doc Ready
+
+//GET List function
+function getList(){
+  clearDom();
+  $.ajax({
+    type: "GET",
+    url: "/list",
+    success: function(response){
+      console.log("getList called");
+      console.log(response);
+      var list = response;
+      for (var i = 0; i < response.length; i++) {
+        var $el = $('#listContainer').children().last();
+        console.log(list[i].id);
+        console.log(list[i].description);
+        console.log(list[i].completed);
+        $el.append("<div data-completed='"+ list[i].completed + "' data-id='" + list[i].id + "'><p>" + list[i].description + "</p><button class='deleteButton'>Delete</button><button class='completeButton'>Complete</button></div>"
+          // "<div class='listDiv' id='listDiv" + i +
+          //           "'<button class='deleteButton' id='deleteButton" + i +
+          //           "'>Delete</button><button class='completeButton' id='completeButton" + i +
+          //           "'>Complete</button></div"
+                  );
+       } //end for loop
+    }// end success
+  }); //end ajax
+} //end getList
 
 //put function to change completed from false to true
 function completed(div){
